Beginning to work on the physics system now. Creating transform component so there is a unified source of where the object is located both for drawing and moving

+++ Ebony/src/components/rigidbodyComponent.hpp
@@ -63,7 +63,7 @@ namespace components
 
 		std::uint16_t getNextScriptedMovementLength()
 		{
-			return scriptedMovements.size();
+			return static_cast<std::uint16_t>(scriptedMovements.size());
 		}
 
 		// Returns the next force and removes it from the list
@@ -83,7 +83,7 @@ namespace components
 
 		std::uint16_t getForceLength()
 		{
-			return forces.size();
+			return static_cast<std::uint16_t>(forces.size());
 		}
